WIP: Implement function to extract sv_flags from stat for macOS
Still need to use this output with some constant to infer whether this
file is hidden or not

This function (_sv_flags) could definitely be improved.  One way would
be to dynamically allocate required size, but I also hate that we are
indexing stat buffer directly.  The sv_flags field should always be in
position 11 (or 21 if using 32-bit), but I still don't like it

Addresses #1