Finish implementing UF_HIDDEN check on file
After trialing many different methods for parsing the sv_flags output
using UF_HIDDEN (all of which had varying resources), I think the one
that works is the one I have implemented.  This finalises the macOS
implementation of this function.

I still have improvements to add, as specified in d3ff556, but these
can come later.

Addresses #1.  Leaving the issue open in the interest of improvements
and verifying that _isinvisible is correctly defined.