Add parametric comments to C calls
C calls (`ccall`), especially ones with many arguments, are confusing
to read if you are not familiar with the function you are calling.  I
wanted to ensure they are somewhat readable beyond just good variable
names.  I briefly used `@ccall` instead, as that sometimes reads
nicer, however it still just relied on good variable names to tell the
reader what the C call is doing.  I ended up pasting the C (or C# or
Objective-C) function header as a comment above each C call so as to
better document what parameters the C calls are expecting.