Move realpath path expansion call to within _ishidden_unix
We need to expand the path into its real path in order to correctly
assess whether the file begins with a full stop.  However, we only
need to check this for Unix.

We should probably add some tests on macOS and BSD in particular to
see if our custom C calls and low-level functions still work without
realpath at the top-most level, or if we need to call normpath as well.