Add notes on missing functionality to primary docstring
Namely, missing functionality in ZFS and UNIX-specific macOS system
dirs.  I'd love to get around to implementing these but haven't
yet (it's difficult and I haven't found the time; might as well
document them so people are aware of this and consider trying again
when I have time or if somebody needs the functionality).